+++ packages/react-headless/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @opencx/widget-react-headless
 
+## 4.0.39
+
+### Patch Changes
+
+- add Bulgarian, Czech, Thai, and Vietnamese language translations
+- Updated dependencies
+  - @opencx/widget-core@4.0.39
+
 ## 4.0.38
 
 ### Patch Changes